requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-mathfilters==1.0.0
djangorestframework==3.15.2
idna==3.7
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
psycopg2==2.9.9
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
